% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova1f_3c.R
\name{anova1f_3c}
\alias{anova1f_3c}
\title{Compute power for a One Factor ANOVA with three levels and contrasts.
Takes means, sds, and sample sizes for each group. Alpha is .05 by default, alternative values may be entered by user}
\usage{
anova1f_3c(
  m1 = NULL,
  m2 = NULL,
  m3 = NULL,
  s1 = NULL,
  s2 = NULL,
  s3 = NULL,
  n1 = NULL,
  n2 = NULL,
  n3 = NULL,
  alpha = 0.05,
  c1 = 0,
  c2 = 0,
  c3 = 0
)
}
\arguments{
\item{m1}{Mean of first group}

\item{m2}{Mean of second group}

\item{m3}{Mean of third group}

\item{s1}{Standard deviation of first group}

\item{s2}{Standard deviation of second group}

\item{s3}{Standard deviation of third group}

\item{n1}{Sample size for first group}

\item{n2}{Sample size for second group}

\item{n3}{Sample size for third group}

\item{alpha}{Type I error (default is .05)}

\item{c1}{Weight for Contrast 1 (default is 0)}

\item{c2}{Weight for Contrast 2 (default is 0)}

\item{c3}{Weight for Contrast 3 (default is 0)}
}
\value{
Power for the One Factor ANOVA
}
\description{
Compute power for a One Factor ANOVA with three levels and contrasts.
Takes means, sds, and sample sizes for each group. Alpha is .05 by default, alternative values may be entered by user
}
\examples{
anova1f_3c(m1=80, m2=82, m3=82, s1=10, s2=10, s3=10,
n1=60, n2=60, n3=60, c1=2, c2=-1, c3=-1, alpha=.05)
}
