% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Assumptions.R
\name{Assumptions}
\alias{Assumptions}
\title{Compute power for Multiple Regression with Violated assumptions (Beta)}
\usage{
Assumptions(
  ry1 = NULL,
  ry2 = NULL,
  ry3 = NULL,
  ry4 = NULL,
  ry5 = NULL,
  r12 = NULL,
  r13 = NULL,
  r14 = NULL,
  r15 = NULL,
  r23 = NULL,
  r24 = NULL,
  r25 = NULL,
  r34 = NULL,
  r35 = NULL,
  r45 = NULL,
  sy = NULL,
  s1 = NULL,
  s2 = NULL,
  s3 = NULL,
  s4 = NULL,
  s5 = NULL,
  ky = NULL,
  k1 = NULL,
  k2 = NULL,
  k3 = NULL,
  k4 = NULL,
  k5 = NULL,
  n = NULL,
  alpha = 0.05,
  test = NULL,
  nruns = 5000
)
}
\arguments{
\item{ry1}{Correlation between DV (y) and first predictor (1)}

\item{ry2}{Correlation between DV (y) and second predictor (2)}

\item{ry3}{Correlation between DV (y) and third predictor (3)}

\item{ry4}{Correlation between DV (y) and fourth predictor (4)}

\item{ry5}{Correlation between DV (y) and fifth predictor (5)}

\item{r12}{Correlation between first (1) and second predictor (2)}

\item{r13}{Correlation between first (1) and third predictor (3)}

\item{r14}{Correlation between first (1) and fourth predictor (4)}

\item{r15}{Correlation between first (1) and fifth predictor (5)}

\item{r23}{Correlation between second (2) and third predictor (3)}

\item{r24}{Correlation between second (2) and fourth predictor (4)}

\item{r25}{Correlation between second (2) and fifth predictor (5)}

\item{r34}{Correlation between third (3) and fourth predictor (4)}

\item{r35}{Correlation between third (3) and fifth predictor (5)}

\item{r45}{Correlation between fourth (4) and fifth predictor (5)}

\item{sy}{Skew of outcome variable}

\item{s1}{Skew of first predictor}

\item{s2}{Skew of second predictor}

\item{s3}{Skew of third predictor}

\item{s4}{Skew of fourth predictor}

\item{s5}{Skew of fifth predictor}

\item{ky}{Kurtosis of outcome variable}

\item{k1}{Kurtosis of first predictor}

\item{k2}{Kurtosis of second predictor}

\item{k3}{Kurtosis of third predictor}

\item{k4}{Kurtosis of fourth predictor}

\item{k5}{Kurtosis of fifth predictor}

\item{n}{Sample size}

\item{alpha}{Type I error (default is .05)}

\item{test}{type of test (none, sqrt, log, inv, robust, boot, quantile, hc0, hc1, hc2, hc3)}

\item{nruns}{number of tuns, default is 5000

#'@examples
Assumptions(ry1=.0, ry2=.3, ry3=.3, ry4=.1, r12 = .0,
r13=.0, r14=.0, r23=.0, r24=.0,r34=0,
sy=1,s1=2,s2=2, s3=1,s4=1, ky=1,k1=1,k2=1,
k3=1,k4=1, n=100, nruns = 100, test="sqrt")}
}
\value{
Power for Multiple Regression with Non Normal Variables
}
\description{
Compute power for Multiple Regression with Violated assumptions (Beta)
}
