% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.R
\name{corr}
\alias{corr}
\title{Compute power for Pearson's Correlation
Takes correlation and range of values}
\usage{
corr(r, nlow, nhigh, alpha = 0.05, tails = 2, by = 1)
}
\arguments{
\item{r}{Correlation}

\item{nlow}{Starting sample size}

\item{nhigh}{Ending sample size}

\item{alpha}{Type I error (default is .05)}

\item{tails}{one or two-tailed tests (default is 2)}

\item{by}{Incremental increase in sample size from low to high}
}
\value{
Power for Pearson's Correlation
}
\description{
Compute power for Pearson's Correlation
Takes correlation and range of values
}
\examples{
corr(r=.30, nlow=60, nhigh=100,by=2)
}
