% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRcat.R
\name{LRcat}
\alias{LRcat}
\title{Compute Power for Logistic Regression with a Single Categorical Predictor}
\usage{
LRcat(p0 = NULL, p1 = NULL, prop = 0.5, alpha = 0.05, power,
  R2 = 0)
}
\arguments{
\item{p0}{Probability of a Desirable Outcome in the Control Condition}

\item{p1}{Probability of a Desirable Outcome in the Treatment Condition}

\item{prop}{Proportion in the Treatment Condition}

\item{alpha}{Type I error (default is .05)}

\item{power}{Desired Power}

\item{R2}{How Well Predictor of Interest is Explained by Other Predictors (default is 0)}
}
\value{
Power for Logistic Regression with a Single Categorical Predictor
}
\description{
Compute Power for Logistic Regression with a Single Categorical Predictor
}
\examples{
LRcat(p0=.137,p1=.611,prop =.689,power=.95)
}
