% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRcont.R
\name{LRcont}
\alias{LRcont}
\title{Compute Power for Logistic Regression with Continuous Predictors}
\usage{
LRcont(OR = NA, r = NA, ER = NULL, alpha = 0.05, power = NULL, R2 = 0)
}
\arguments{
\item{OR}{Odds Ratio for Predictor of Interest}

\item{r}{Correlation for Predictor of Interest}

\item{ER}{Event Ratio Probability of a Desirable Outcome Overall}

\item{alpha}{Type I error (default is .05)}

\item{power}{Desired Power}

\item{R2}{How Well Predictor of Interest is Explained by Other Predictors (default is 0)}
}
\value{
Power for Logistic Regression with Continuous Predictors
}
\description{
Compute Power for Logistic Regression with Continuous Predictors
}
\examples{
LRcont(OR = 4.05, ER = .463,power=.95)
}
