% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modmed15.R
\name{modmed15}
\alias{modmed15}
\title{Compute Power for Conditional Process Model 15 Joint Significance
Requires correlations between all variables as sample size.
This is the recommended approach for determining power}
\usage{
modmed15(
  rxw,
  rxm,
  rxwm,
  rxy,
  rwm = 0,
  rwwm,
  rwxw,
  rxwwx,
  rwy,
  rmwm = 0,
  rmwx,
  rmwy,
  rwxy,
  rwmwx,
  rmy,
  rxwx,
  rxwy,
  rmwwx,
  n,
  alpha = 0.05,
  rep = 5000
)
}
\arguments{
\item{rxw}{Correlation between predictor (x) and moderator (w)}

\item{rxm}{Correlation between predictor (x) and mediator (m)}

\item{rxwm}{Correlation between predictor (x) and interaction term (xw) - defaults to 0}

\item{rxy}{Correlation between DV (y) and predictor (x)}

\item{rwm}{Correlation between moderator (w) and mediator (m)}

\item{rmwm}{Correlation between mediator (m) and interaction (xw) - Key value}

\item{rmwy}{Correlation between DV (y) and interaction (xw) - defaults to 0}

\item{n}{Sample size}

\item{alpha}{rep Number of samples drawn (defaults to 5000)}

\item{rym}{Correlation between DV (y) and mediator (m)}

\item{ryw}{Correlation between DV (y) and moderator (w)}

\item{rmww}{Correlation between moderator (w) and interaction (xw) - defaults to 0}
}
\value{
Power for Model 7 Conditional Processes
}
\description{
Compute Power for Conditional Process Model 15 Joint Significance
Requires correlations between all variables as sample size.
This is the recommended approach for determining power
}
\examples{
modmed15(rxw=.40, rxm=.4, rxwm=.0, rxy=.5, rwm=.45, rmww=.0, rwy=.2,
rmwm=.45,rmy=.30,rmwy=.0,rep=5000, alpha=.05, n=400)
}
