% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modmed7.R
\name{modmed7}
\alias{modmed7}
\title{Compute Power for Model 7 Conditional Processes Using Joint Significance
Requires correlations between all variables as sample size.
This is the recommended approach for determining power}
\usage{
modmed7(
  rxm,
  rxw,
  rxxw,
  rxy,
  rwm,
  rwxw,
  rwy = 0,
  rmxw,
  rmy,
  rxwy = 0,
  alpha = 0.05,
  rep = 1000,
  n = NULL
)
}
\arguments{
\item{rxm}{Correlation between predictor (x) and mediator (m)}

\item{rxw}{Correlation between predictor (x) and moderator (w)}

\item{rxxw}{Correlation between predictor (x) and interaction term (xw) - defaults to 0}

\item{rxy}{Correlation between DV (y) and predictor (x)}

\item{rwm}{Correlation between moderator (w) and mediator (m)}

\item{rwxw}{Correlation between moderator (w) and interaction (xw) - defaults to 0}

\item{rmxw}{Correlation between mediator (m) and interaction (xw) - Key value}

\item{rxwy}{Correlation between DV (y) and interaction (xw) - defaults to 0}

\item{alpha}{rep Number of samples drawn (defaults to 5000)}

\item{n}{Sample size}

\item{rym}{Correlation between DV (y) and mediator (m)}

\item{ryw}{Correlation between DV (y) and moderator (w)}
}
\value{
Power for Model 7 Conditional Processes
}
\description{
Compute Power for Model 7 Conditional Processes Using Joint Significance
Requires correlations between all variables as sample size.
This is the recommended approach for determining power
}
\examples{
medmod7(rxm=.4, rxw=.4, rxxw=.0, rxy=.50,rxwm=.45, rmy=.3, rxwy=0,rwm=.45,
rxww=0,rwy=.2,alpha=.05,rep=1000,n=400)
}
