% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modmed8.R
\name{modmed8}
\alias{modmed8}
\title{Compute Power for Power for Model 8 Conditional Processes Using Joint Significance
Requires correlations between all variables as sample size.
This is the recommended approach for determining power}
\usage{
modmed8(
  rxw,
  rxm,
  rxxw,
  rxy,
  rwm = 0,
  rxww,
  rwy,
  rxwm,
  rxwy,
  rwxw,
  rmy,
  n,
  alpha = 0.05,
  rep = 5000
)
}
\arguments{
\item{rxw}{Correlation between predictor (x) and moderator (w)}

\item{rxm}{Correlation between predictor (x) and mediator (m)}

\item{rxxw}{Correlation between predictor (x) and interaction term (xw) - defaults to 0}

\item{rxy}{Correlation between DV (y) and predictor (x)}

\item{rwm}{Correlation between moderator (w) and mediator (m)}

\item{rxww}{Correlation between moderator (w) and interaction (xw) - defaults to 0}

\item{rxwm}{Correlation between mediator (m) and interaction (xw) - Key value}

\item{rxwy}{Correlation between DV (y) and interaction (xw) - defaults to 0}

\item{n}{Sample size}

\item{alpha}{rep Number of samples drawn (defaults to 5000)}

\item{rym}{Correlation between DV (y) and mediator (m)}

\item{ryw}{Correlation between DV (y) and moderator (w)}
}
\value{
Power for Model 7 Conditional Processes
}
\description{
Compute Power for Power for Model 8 Conditional Processes Using Joint Significance
Requires correlations between all variables as sample size.
This is the recommended approach for determining power
}
\examples{
modmed8(rxw<-.2, rxm<-.3, rxxw=0, rxy=.3,rwm=.4,
rxww=0,rwy<-.35, rxwm<-.4, rxwy=0, rmy=.3, n=100, rep=2000,alpha=.05)
}
