% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRC.R
\name{MRC}
\alias{MRC}
\title{Compute power for Multiple Regression with up to Five Predictors
Example code below for three predictors. Expand as needed for four or five}
\usage{
MRC(ry1 = NULL, ry2 = NULL, ry3 = NULL, ry4 = NULL, ry5 = NULL,
  r12 = NULL, r13 = NULL, r14 = NULL, r15 = NULL, r23 = NULL,
  r24 = NULL, r25 = NULL, r34 = NULL, r35 = NULL, r45 = NULL,
  n = NULL, alpha = 0.05, rep = 10000, my = 0, m1 = 0, m2 = 0,
  m3 = 0, m4 = 0, m5 = 0, sy = 1, s1 = 1, s2 = 1, s3 = 1,
  s4 = 1, s5 = 1)
}
\arguments{
\item{ry1}{Correlation between DV (y) and first predictor (1)}

\item{ry2}{Correlation between DV (y) and second predictor (2)}

\item{ry3}{Correlation between DV (y) and third predictor (3)}

\item{ry4}{Correlation between DV (y) and fourth predictor (4)}

\item{ry5}{Correlation between DV (y) and fifth predictor (5)}

\item{r12}{Correlation between first (1) and second predictor (2)}

\item{r13}{Correlation between first (1) and third predictor (3)}

\item{r14}{Correlation between first (1) and fourth predictor (4)}

\item{r15}{Correlation between first (1) and fifth predictor (5)}

\item{r23}{Correlation between second (2) and third predictor (3)}

\item{r24}{Correlation between second (2) and fourth predictor (4)}

\item{r25}{Correlation between second (2) and fifth predictor (5)}

\item{r34}{Correlation between third (3) and fourth predictor (4)}

\item{r35}{Correlation between third (3) and fifth predictor (5)}

\item{r45}{Correlation between fourth (4) and fifth predictor (5)}

\item{n}{Sample size}

\item{alpha}{Type I error (default is .05)}

\item{rep}{number of replications (default is 10000)}

\item{my}{Mean of DV (default is 0)}

\item{m1}{Mean of first predictor (default is 0)}

\item{m2}{Mean of second redictor (default is 0)}

\item{m3}{Mean of third predictor (default is 0)}

\item{m4}{Mean of fourth predictor (default is 0)}

\item{m5}{Mean of fifth predictor (default is 0)}

\item{sy}{Standard deviation of DV (default is 1)}

\item{s1}{Standard deviation of first predictor (default is 1)}

\item{s2}{Standard deviation of second predictor (default is 1)}

\item{s3}{Standard deviation of third predictor (default is 1)}

\item{s4}{Standard deviation of fourth predictor (default is 1)}

\item{s5}{Standard deviation of fifth predictor (default is 1)}
}
\value{
Power for Multiple Regression with Two to Five Predictors
}
\description{
Compute power for Multiple Regression with up to Five Predictors
Example code below for three predictors. Expand as needed for four or five
}
