% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRC_all.R
\name{MRC_all}
\alias{MRC_all}
\title{Compute power for Mutliple Regression with Three Predictors
Requires correlations between all variables as sample size. Means, sds, and alpha are option. Also computes Power(All)}
\usage{
MRC_all(ry1 = NULL, ry2 = NULL, ry3 = NULL, r12 = NULL,
  r13 = NULL, r23 = NULL, n = NULL, alpha = 0.05, rep = 10000,
  my = 0, m1 = 0, m2 = 0, m3 = 0, sy = 1, s1 = 1, s2 = 1,
  s3 = 1)
}
\arguments{
\item{ry1}{Correlation between DV (y) and first predictor (1)}

\item{ry2}{Correlation between DV (y) and second predictor (2)}

\item{ry3}{Correlation between DV (y) and third predictor (3)}

\item{r12}{Correlation between first (1) and second predictor (2)}

\item{r13}{Correlation between first (1) and third predictor (3)}

\item{r23}{Correlation between second (2) and third predictor (3)}

\item{n}{Sample size}

\item{alpha}{Type I error (default is .05)}

\item{rep}{number of replications (default is 10000)}
}
\value{
Power for Multiple Regression (ALL)
}
\description{
Compute power for Mutliple Regression with Three Predictors
Requires correlations between all variables as sample size. Means, sds, and alpha are option. Also computes Power(All)
}
