% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairt.R
\name{pairt}
\alias{pairt}
\title{Compute power for a Paired t-test
Takes means, sd, and sample sizes. Alpha is .05 by default, alternative values may be entered by user.
correlation (r) defaults to .50.}
\usage{
pairt(m1 = NULL, m2 = NULL, s = NULL, n = NULL, r = NULL,
  alpha = 0.05)
}
\arguments{
\item{m1}{Mean for Pre Test}

\item{m2}{Mean for Post Test}

\item{s}{Standard deviation}

\item{n}{Sample size}

\item{r}{Correlation pre-post measures (default is .50)}

\item{alpha}{Type I error (default is .05)}
}
\value{
Power for the Paired t-test
}
\description{
Compute power for a Paired t-test
Takes means, sd, and sample sizes. Alpha is .05 by default, alternative values may be entered by user.
correlation (r) defaults to .50.
}
\examples{
pairt(m1=25,m2=20, s = 5, n = 25, r = .5)
}
