% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_prec.R
\name{r_prec}
\alias{r_prec}
\title{Compute Precision Analyses for Correlations
This approach simply loops a function from MBESS}
\usage{
r_prec(r, nlow, nhigh, ci = 0.95, by = 1)
}
\arguments{
\item{r}{Correlation}

\item{nlow}{starting sample size}

\item{nhigh}{ending sample size}

\item{ci}{Type of Confidence Interval (e.g., .95)}

\item{by}{Incremental increase in sample (e.g. nlow = 10, nhigh = 24, by = 2, produces estimates of 10, 12, and 14)}
}
\value{
Precision Analyses for Correlations
}
\description{
Compute Precision Analyses for Correlations
This approach simply loops a function from MBESS
}
\examples{
r_prec(r=.3, nlow=80, nhigh=400, by=20, ci=.95)
}
